+++ BLL/Services/ProfileService.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using BLL.Interfacies.Entities;
 using BLL.Interfacies.Services;
@@ -21,12 +22,20 @@ namespace BLL.Services
 
         public ProfileEntity GetById(int id)
         {
-            return profileRepository.GetById(id).ToBllProfile();
+            if (id <= 0)
+                throw new ArgumentOutOfRangeException();
+            var profile = profileRepository.GetById(id);
+            if (ReferenceEquals(profile, null))
+                return null;
+            return profile.ToBllProfile();
         }
 
         public ProfileEntity GetProfileByName(string name)
         {
-            return profileRepository.GetProfileByName(name).ToBllProfile();
+            var profile = profileRepository.GetProfileByName(name);
+            if (ReferenceEquals(profile, null))
+                return null;
+            return profile.ToBllProfile();
         }
 
         public IEnumerable<ProfileEntity> GetAllEntities()
@@ -36,36 +45,27 @@ namespace BLL.Services
 
         public void Create(ProfileEntity profile)
         {
+            if (ReferenceEquals(profile, null))
+                throw new ArgumentNullException();
+
             profileRepository.Create(profile.ToDalProfile());
-            uow.Commit();
-        }
-
-        public void DeletePhoto(ProfileEntity entity, int id)
-        {
-            profileRepository.DeletePhoto(entity.ToDalProfile(), id);
-            uow.Commit();
-        }
-
-        public void RemoveLike(ProfileEntity profile, LikeEntity likeEntity)
-        {
-            profileRepository.RemoveLike(profile.ToDalProfile(), likeEntity.ToDalLike());
-            uow.Commit();
-        }
-
-        public void AddLike(ProfileEntity profile, LikeEntity likeEntity)
-        {
-            profileRepository.AddLike(profile.ToDalProfile(), likeEntity.ToDalLike());
             uow.Commit();
         }
 
         public void Delete(ProfileEntity profile)
         {
+            if (ReferenceEquals(profile, null))
+                throw new ArgumentNullException();
+
             profileRepository.Delete(profile.ToDalProfile());
             uow.Commit();
         }
 
         public void Update(ProfileEntity item)
         {
+            if (ReferenceEquals(item, null))
+                throw new ArgumentNullException();
+
             profileRepository.Update(item.ToDalProfile());
             uow.Commit();
         }
